Fix cmd->pipeline layering inversion: move colour constants + LiveRegion to watchdog.terminal
pipeline/orchestrate.py and pipeline/preprocess_batch.py imported the six colour
constants from cmd.base and LiveRegion from cmd.live, inverting the natural
CLI-calls-pipeline dependency direction (cmd/base.py itself imports from
pipeline/write_vault.py). Neither cmd.base's colour gating (_color_enabled,
#499/D174) nor cmd/live.py's LiveRegion actually depend on anything else in
cmd/ — live.py in particular has zero imports from the package it sits in — so
both move wholesale, unmodified in behavior, into a new neutral module,
watchdog/terminal.py.

cmd/base.py and cmd/live.py now re-export from watchdog.terminal, so the ~40
existing `from watchdog.cmd.base import _BOLD, ...` call sites across cmd/*.py,
plus tests/test_live.py's `watchdog.cmd.live.shutil.get_terminal_size`
monkeypatch, keep working unchanged. Renamed live.py's own unconditional
`_RESET` (used only to close truncated ANSI spans, deliberately independent of
the colour gate per D174) to `_LIVE_RESET` since both constants now live in
the same module and would otherwise silently collide.

Verified: full suite green, ruff clean, and a manual check that `watchdog
--help` still renders colour on a real TTY (via a pty) and stays plain when
piped.
